+++ src/app/components/Gallery/Gallery.tsx
@@ -5,10 +5,8 @@ import { GalleryImage } from '../GalleryImage/GalleryImage'
 import { GalleryImagePreview } from '../GalleryImagePreview/GalleryImagePreview'
 import { Modal } from '@/components/ui/Modal/Modal'
 import { ImageCarousel } from '@/components/ui/ImageCarousel/ImageCarousel'
-import { useMediaQuery } from '@/helpers/hooks/useMediaQuery'
-import { useGalleryImages } from './hooks/useGalleryImages'
-import { useGalleryImagePreview } from './hooks/useGalleryImagePreview'
-import { useFullscreenImageView } from '@/helpers/hooks/useFullscreenImageView'
+import { useFullscreenImageView, useMediaQuery } from '@/helpers/hooks'
+import { useGalleryImagePreview, useGalleryImages } from './hooks'
 import { TImage } from '@/helpers/types'
 import styles from './Gallery.module.css'
