+++ src/app/components/Gallery/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useGalleryImagePreview'
+export * from './useGalleryImages'
