+++ src/app/components/GalleryImage/GalleryImage.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image'
 import { LightboxButton } from '@/components/ui/LightboxButton/LightboxButton'
-import { useMediaQuery } from '@/helpers/hooks/useMediaQuery'
+import { useMediaQuery } from '@/helpers/hooks'
 import { TImage } from '@/helpers/types'
 import styles from './GalleryImage.module.css'
