+++ src/app/components/Merch/Merch.tsx
@@ -5,8 +5,7 @@ import { List } from '@/components/ui/List/List'
 import { Carousel } from '@/components/ui/Carousel/Carousel'
 import { MerchCard } from '../MerchCard/MerchCard'
 import { Modal } from '@/components/ui/Modal/Modal'
-import { useMediaQuery } from '@/helpers/hooks/useMediaQuery'
-import { useFullscreenImageView } from '@/helpers/hooks/useFullscreenImageView'
+import { useFullscreenImageView, useMediaQuery } from '@/helpers/hooks'
 import { IMerchItem } from '@/helpers/types'
 import styles from './Merch.module.css'
