+++ src/app/components/Music/Music.tsx
@@ -4,8 +4,7 @@ import { List } from '@/components/ui/List/List'
 import { Carousel } from '@/components/ui/Carousel/Carousel'
 import { AlbumItem } from '../AlbumItem/AlbumItem'
 import { AudioTrack } from '@/components/ui/AudioTrack/AudioTrack'
-import { useInitAudioTrackRefs } from './hooks/useInitAudioTrackRefs'
-import { useMusicPlayer } from './hooks/useMusicPlayer'
+import { useInitAudioTrackRefs, useMusicPlayer } from './hooks'
 import { AudioTrackRefsProvider } from '@/helpers/contexts/AudioTrackRefs.context'
 import { ALBUM_CAROUSEL_THRESHOLD } from '@/helpers/constants'
 import { IAlbum } from '@/helpers/types'
