+++ src/app/components/Music/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useInitAudioTrackRefs'
+export * from './useMusicPlayer'
